LibJS: Allow the parser to defer parsing nodes in case of deep nesting
With these changes, the parser can decide when to stop recursively
parsing nested code and return an unparsed thunk for that specific
region of code.
The region to skip is determined by scanning tokens until none of the
following rules match:
- must be able to parse an expression
- failing that, if the previous token was part of an expression, must be
  able to parse a secondary expression
- failing that, we must have balanced enclosures (parens, braces,
  brackets and conditional expressions)

Fixes #11705.
Fixes #12073.
And probably a few clusterfuzz stack overflow issues.