Shell: Do not assume that all string values are valid UTF-8
This assumption was blindly added when switching to the new String, but
it makes little sense to require all strings to be utf-8, as shells have
to deal with all sorts of non-utf-8 strings. This also applies to
substitutions.

Proper names in the shell *are* still subject to the UTF-8 requirements,
so a variable with a silly name will still be invalid.